refactor pt 7: split schemes.py god-module into a package
- split interfaces/dynamics_blending/schemes.py (793 LOC) into schemes/{blending,comp_psinc,swe_lake,hydro_nonhydro,orchestration}.py
- blending: the Blend interface; comp_psinc / swe_lake / hydro_nonhydro: the regime conversions; orchestration: the per-timestep blending calls
- hydro_nonhydro keeps the ~240-line commented do_hydro_to_nonhydro_conv body verbatim as the hydrostatic-blending reinstatement reference
- re-export all public names from schemes/__init__.py so time_update.py and prepare.py call sites are unchanged
- mechanical move (verbatim ranges, only relative-import depth fixed): FULL repro gate bit-identical at tol 0 across all 11 golden masters
